backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "G-Tool - 메일, 캘린더, 할일, 북마크 통합 생산성 플랫폼"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "sqlalchemy",
    "aiosqlite",
    "python-dotenv",
    "httpx",
    "openai",
    "pydantic-settings",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "apscheduler>=3.10",
    "PyJWT>=2.8",
    "cryptography>=43.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
